stackredis/python-app/requirements.txt
flask
redis
validators
cassandra-driver
